Develop

+++ src/app/presentation/features/login/login.component.ts
@@ -1,22 +1,25 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { error } from 'console';
 import { Observable } from 'rxjs';
 import { PerfilInversorAPI } from 'src/app/core/models/API/Perfil-Inversor-API.model';
 import { UsuarioAPI } from 'src/app/core/models/API/Usuario-API.model';
 import { LocalStorageService } from 'src/app/core/services/LocalStorage/local-storage.service';
 import { RegistroService } from 'src/app/core/services/api/autorizacion/registro.service';
 import { PerfilInversorService } from 'src/app/core/services/api/perfil-inversor/perfil-inversor.service';
+import { jwtDecode } from "jwt-decode";
+
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.scss']
 })
-export class LoginComponent implements OnInit {
-
+export class LoginComponent{
+  opcion = 'login';
   errorLogin: boolean = false;
   errorform: boolean = false;
+  errorRecuperar = false;
+  successRecuperar = false;
   usuarioForm: any = {
     email: '',
     password: ''
@@ -42,10 +45,24 @@ export class LoginComponent implements OnInit {
     private router: Router
   ) { }
 
-  ngOnInit(): void {
+  public opcionRecuperar() {
+    this.opcion = 'recuperar';
+  }
+  public recuperarCuenta() {
+    this.loading = true;
+
+    this.registroUsuarioService.RecuperarCuenta(this.usuarioForm.email).subscribe(
+      (data) => {
+        this.loading = false;
+        this.errorRecuperar = false;
+        this.successRecuperar = true;
+      },
+      (error) => {
+        this.loading = false;
+        this.errorRecuperar = true;
+      });
 
   }
-
   public loguearme() {
 
     this.errorform = false;
@@ -58,7 +75,6 @@ export class LoginComponent implements OnInit {
     }
 
     this.loginUsuario();
-    // this.verfificarUsuario();
 
   }
 
@@ -68,11 +84,18 @@ export class LoginComponent implements OnInit {
     this.registroUsuarioService.loginUsuario(this.usuarioForm.email, this.usuarioForm.password).subscribe(
       (data) => {
         this.LocalStorageService.setItem("token", data.token);
+        const tokenDecoded: any = { ...jwtDecode(data.token) };
+        if (tokenDecoded.esAdministrador) {
+          this.loginAdministrador();
+          return
+        }
+
         this.registroUsuarioService.buscarUsuario(this.usuarioForm.email).subscribe(
           (usuarioRecibido: UsuarioAPI) => {
             this.LocalStorageService.setUsuarioPerfilActualLocal(null);
             this.LocalStorageService.RemovePerfilActualLocal();
             this.LocalStorageService.removeItem('Username');
+            this.LocalStorageService.removeItem('perfilObjetivoCartera')
             this.usuariodb = usuarioRecibido;
             this.buscarPerfilUsuario(this.usuariodb).subscribe(
               (perfilUsuario: PerfilInversorAPI | null) => {
@@ -84,7 +107,6 @@ export class LoginComponent implements OnInit {
             this.loading = false;
             this.errorform = false;
             this.errorLogin = true;
-            console.log("🚀 ~ file: login.component.ts:82 ~ LoginComponent ~ verfificarUsuario ~ errorLogin:", this.errorLogin)
             console.error("Error al buscar Usuario", error);
           });
       },
@@ -92,42 +114,17 @@ export class LoginComponent implements OnInit {
         this.loading = false;
         this.errorform = false;
         this.errorLogin = true;
-        console.log("🚀 ~ file: login.component.ts:82 ~ LoginComponent ~ verfificarUsuario ~ errorLogin:", this.errorLogin)
         console.error("Error al buscar Usuario", error);
       });
   }
 
-
-  public verfificarUsuario() {
-
-    console.log("Verificar Usuario");
-    // this.loading = true;
-
-    // this.registroUsuarioService.buscarUsuario(this.usuarioForm.email).subscribe(
-    //   (usuarioRecibido: UsuarioAPI) => {
-    //     this.LocalStorageService.setUsuarioPerfilActualLocal(null);
-    //     this.LocalStorageService.RemovePerfilActualLocal();
-    //     this.LocalStorageService.removeItem('Username');
-    //     this.usuariodb = usuarioRecibido;
-    //     this.buscarPerfilUsuario(this.usuariodb).subscribe(
-    //       (perfilUsuario: PerfilInversorAPI | null) => {
-    //         this.AlmacenarUsuario_Perfil(perfilUsuario);
-    //       }
-    //     );
-    //   },
-    //   (error) => {
-    //     this.loading = false;
-    //     this.errorform = false;
-    //     this.errorLogin = true;
-    //     console.log("🚀 ~ file: login.component.ts:82 ~ LoginComponent ~ verfificarUsuario ~ errorLogin:", this.errorLogin)
-    //     console.error("Error al buscar Usuario", error);
-    //   });
+  public loginAdministrador() {
+    this.loading = false;
+    this.navegarAHomeAdministrador();
   }
 
   private AlmacenarUsuario_Perfil(perfilUsuario: PerfilInversorAPI | null) {
-    //A nivel login la variable se llama username
     this.LocalStorageService.setItem('Username', this.usuariodb.username);
-    console.log("🚀 ~ file: login.component.ts:78 ~ LoginComponent ~ verfificarUsuario ~ this.usuariodb.nombreUsuario:", this.usuariodb.username)
     if (perfilUsuario === null || perfilUsuario === undefined) {
       console.log("El usuario no tiene un perfil asociado");
       this.loading = false;
@@ -139,14 +136,17 @@ export class LoginComponent implements OnInit {
       this.loading = false;
       this.LocalStorageService.setPerfilSubjetivo(perfilUsuario);
       this.LocalStorageService.SetPerfilActualLocal();
+      this.LocalStorageService.setItem('perfilinversor', perfilUsuario.perfilInversor);
+      if(perfilUsuario.tipoNivelConocimiento!=null||perfilUsuario.nivelConocimiento!=null){
+      this.LocalStorageService.setItem('perfilObjetivoCartera', perfilUsuario.perfilInversor);
+      }
+     
       this.navegarAHome();
     }
   }
 
   buscarPerfilUsuario(usuariodb: UsuarioAPI): Observable<PerfilInversorAPI | null> {
-
     return this.perfilesServicios.obtenerPerfil(usuariodb)
-
   }
 
   validarEntradas() {
@@ -161,6 +161,10 @@ export class LoginComponent implements OnInit {
 
   navegarAHome() {
     this.router.navigate(['dashboard/cartera']);
+  }
+
+  navegarAHomeAdministrador() {
+    this.router.navigate(['dashboard/administrar']);
   }
 
 }
